refactor: separate TodoList state logics to useTodoState hook
