+++ src/pages/TodoPage.tsx
@@ -1,78 +1,25 @@
 import { Navigate } from 'react-router-dom'
 import { useAuthState } from '../AuthProvider'
 import styled from '@emotion/styled'
-import { useEffect, useState } from 'react'
-import { TodoType, createTodo, deleteTodo, getTodo, updateTodo } from './api'
+import { useEffect } from 'react'
 import Todo from '../components/Todo'
+import useTodoState from '../hooks/useTodoState'
 
 const TodoPage = () => {
   const { userState } = useAuthState()
-  const [todos, setTodos] = useState<TodoType[]>([])
-  const [isLoading, setIsLoading] = useState(true)
-  const [isError, setIsError] = useState('')
-  const [newTodo, setNewTodo] = useState('')
-  const [focusedTodoId, setFocusedTodoId] = useState<number | null>(null)
-
-  const getTodoRequest = async () => {
-    const res = await getTodo()
-    if (res.statusCode === 200) {
-      setTodos(() => res.body as TodoType[])
-    } else {
-      setIsError(res.body as string)
-    }
-    setIsLoading(false)
-  }
-
-  const createTodoRequest = async (todo: string) => {
-    if (!todo.length) {
-      alert('Todo string is empty!')
-      setIsLoading(false)
-      return
-    }
-    const res = await createTodo(todo)
-    if (res.statusCode === 201) {
-      setTodos((prev) => [...prev].concat(res.body as TodoType[]))
-      setNewTodo('')
-    } else {
-      setIsError(res.body as string)
-    }
-    setIsLoading(false)
-  }
-
-  const updateTodoRequest = async (
-    todoId: number,
-    todo: string,
-    isCompleted: boolean
-  ) => {
-    const res = await updateTodo(todoId, todo, isCompleted)
-    if (res.statusCode === 200) {
-      const updatedTodo = res.body[0] as TodoType
-      setTodos((prev) =>
-        [...prev].map((todo) => {
-          if (todo.id === updatedTodo.id) {
-            return { ...todo, ...updatedTodo }
-          }
-          return todo
-        })
-      )
-    } else {
-      setIsError(res.body as string)
-    }
-    setIsLoading(false)
-  }
-
-  const deleteTodoRequest = async (todoId: number) => {
-    const res = await deleteTodo(todoId)
-    if (res.statusCode === 204) {
-      setTodos((prev) => [...prev].filter((todo) => todo.id !== todoId))
-    } else {
-      setIsError(res.body as string)
-    }
-    setIsLoading(false)
-  }
+  const {
+    todos,
+    isLoading,
+    isError,
+    newTodo,
+    setNewTodo,
+    focusedTodoId,
+    setFocusedTodoId,
+    apiRequests,
+  } = useTodoState()
 
   useEffect(() => {
-    getTodoRequest()
+    userState.auth && apiRequests.getTodoRequest()
   }, [])
 
   if (isLoading) return <div>Loading</div>
@@ -89,13 +36,13 @@ const TodoPage = () => {
           onChange={(e) => setNewTodo(() => e.target.value)}
           onKeyDown={(e) => {
             if (e.key === 'Enter') {
-              createTodoRequest(newTodo)
+              apiRequests.createTodoRequest(newTodo)
             }
           }}
         />
         <button
           data-testid="new-todo-add-button"
-          onClick={() => createTodoRequest(newTodo)}
+          onClick={() => apiRequests.createTodoRequest(newTodo)}
         >
           추가
         </button>
@@ -108,8 +55,8 @@ const TodoPage = () => {
               text={todo}
               completed={isCompleted}
               isEditable={focusedTodoId === id}
-              handleUpdate={updateTodoRequest}
-              handleDelete={deleteTodoRequest}
+              handleUpdate={apiRequests.updateTodoRequest}
+              handleDelete={apiRequests.deleteTodoRequest}
               handleEditable={(id: number | null) => {
                 setFocusedTodoId(id)
               }}
